refactor: flatten topology
