+++ src/WebClient/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
+    "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.tsbuildinfo",
     "target": "ESNext",
     "useDefineForClassFields": true,
     "lib": ["ES2020", "DOM", "DOM.Iterable"],
@@ -18,7 +18,7 @@
     "allowJs": true,
     "incremental": true,
     "paths": {
-      "@dotcom/*": ["./app/*"]
+      "@dotcom/*": ["./*"]
     },
 
     /* Linting */
@@ -36,8 +36,8 @@
     "resolveJsonModule": true
   },
   "include": [
-    "app/**/*.ts",
-    "app/**/*.tsx",
+    "**/*.ts",
+    "**/*.tsx",
     "globals.d.ts",
     "next-env.d.ts",
     ".next/types/**/*.ts"
