+++ src/WebClient/types/index.ts
@@ -1,4 +1,4 @@
-export * from '@dotcom/types/certification';
+export * from '@dotcom/types/certificate';
 export * from '@dotcom/types/company';
 export * from '@dotcom/types/school';
 export * from '@dotcom/types/job';
