+++ src/WebClient/types/school.ts
